+++ R/registration.R
@@ -1886,9 +1886,7 @@ transferParameterInput <- function(params, image_list) {
 
 .ALIGNMENT_ACCURACY_METRICS <- c(
   "SSIM",
-  "Matte's MI",
-  "Intersection",
-  "Bhattacharyya"
+  "Matte's MI"
 )
 
 .ALIGNMENT_KEYPOINT_METRICS <- c(
@@ -3999,7 +3997,7 @@ getSimpleITKAutomatedRegistration <- function(
   elx$SetParameterMap(mp)
   elx$LogToConsoleOff()
   tmp <- elx$Execute()
-  sitk_img <- SimpleITK::ReadImage(file.path(tmpdir, "result.0.tif"))
+  sitk_img <- elx$GetResultImage()
   transform_param_map <- elx$GetTransformParameterMap()
   tfx_image <- SimpleITK::TransformixImageFilter()
   tfx_image$LogToConsoleOff()